requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.13
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
apscheduler==3.11.0
    # via -r requirements.in
attrs==25.1.0
    # via aiohttp
boto3==1.37.3
    # via
    #   -r requirements.in
    #   langchain-aws
boto3-stubs==1.37.3
    # via -r requirements.in
botocore==1.37.3
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.37.1
    # via
    #   boto3-stubs
    #   types-boto3
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.9
    # via -r requirements.in
fastapi-cli==0.0.7
    # via fastapi
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.24.1
    # via google-api-python-client
google-api-python-client==2.162.0
    # via
    #   -r requirements.in
    #   google-api-python-client-stubs
google-api-python-client-stubs==1.29.0
    # via -r requirements.in
google-auth==2.38.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
googleapis-common-protos==1.68.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   langsmith
    #   notion-client
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.5
    # via fastapi
jiter==0.8.2
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.19
    # via
    #   -r requirements.in
    #   langchain-community
langchain-aws==0.2.14
    # via -r requirements.in
langchain-community==0.3.18
    # via -r requirements.in
langchain-core==0.3.40
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-aws
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.3.6
    # via langchain
langsmith==0.3.11
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown==3.7
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
notion-client==2.3.0
    # via -r requirements.in
numpy==1.26.4
    # via
    #   langchain
    #   langchain-aws
    #   langchain-community
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.65.1
    # via -r requirements.in
orjson==3.10.15
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pillow==11.1.0
    # via -r requirements.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.0
    # via google-api-core
protobuf==5.29.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-aws
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via langchain-community
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pytz==2025.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via
    #   -r requirements.in
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
rsa==4.9
    # via google-auth
s3transfer==0.11.3
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.38
    # via
    #   langchain
    #   langchain-community
    #   sqlmodel
sqlmodel==0.0.22
    # via -r requirements.in
starlette==0.45.3
    # via fastapi
structlog==25.1.0
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tqdm==4.67.1
    # via openai
typer==0.15.2
    # via fastapi-cli
types-awscrt==0.23.10
    # via botocore-stubs
types-boto3==1.37.3
    # via -r requirements.in
types-boto3-bedrock==1.37.0
    # via types-boto3
types-boto3-cloudformation==1.37.0
    # via types-boto3
types-boto3-dynamodb==1.37.0
    # via types-boto3
types-boto3-ec2==1.37.2
    # via types-boto3
types-boto3-lambda==1.37.0
    # via types-boto3
types-boto3-rds==1.37.0
    # via types-boto3
types-boto3-s3==1.37.0
    # via types-boto3
types-boto3-secretsmanager==1.37.0
    # via types-boto3
types-boto3-sqs==1.37.0
    # via types-boto3
types-httplib2==0.22.0.20241221
    # via google-api-python-client-stubs
types-s3transfer==0.11.3
    # via
    #   boto3-stubs
    #   types-boto3
typing-extensions==4.12.2
    # via
    #   anyio
    #   boto3-stubs
    #   fastapi
    #   google-api-python-client-stubs
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
    #   types-boto3
    #   types-boto3-bedrock
    #   types-boto3-cloudformation
    #   types-boto3-dynamodb
    #   types-boto3-ec2
    #   types-boto3-lambda
    #   types-boto3-rds
    #   types-boto3-s3
    #   types-boto3-secretsmanager
    #   types-boto3-sqs
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzlocal==5.3
    # via apscheduler
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.3.0
    # via
    #   botocore
    #   requests
uvicorn==0.34.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0
    # via uvicorn
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith
